Исправила код согласно комментариев наставника по модулям 5-7
