+++ js/create-photos.js
@@ -1,5 +1,24 @@
 import {getRandomInteger, createRandomIdFromRangeGenerator, getRandomArrayElement} from './utils.js';
-import {getArraysWithMessagesAndNames} from './arrays-with-messages-and-names.js';
+
+const messages = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const names = [
+  'Вася',
+  'Коля',
+  'Петя',
+  'Таня',
+  'Ира',
+  'Дина',
+  'Ваня',
+  'Маша'
+];
 
 const PHOTO_ID_MIN = 1;
 const PHOTO_ID_MAX = 25;
@@ -7,10 +26,10 @@ const LIKES_MIN = 15;
 const LIKES_MAX = 200;
 const COMMENTS_COUNT_MIN = 0;
 const COMMENTS_COUNT_MAX = 30;
-const DESCRIPTIONS_COUNT = 25;
+
 const AVATAR_ID_MIN = 1;
 const AVATAR_ID_MAX = 6;
-const {messages, names} = getArraysWithMessagesAndNames();
+//const {messages, names} = getArraysWithMessagesAndNames();
 
 const createIdGenerator = (min, max) => {
   let lastGeneratedId = min;
@@ -43,6 +62,6 @@ const createPhoto = () => ({
   comments: Array.from({length:getRandomInteger(COMMENTS_COUNT_MIN,COMMENTS_COUNT_MAX)}, createComment),
 });
 
-const photos = Array.from({length: DESCRIPTIONS_COUNT}, createPhoto);
+const createPhotos = (size) => Array.from({length: size}, createPhoto);
 
-export {photos};
+export {createPhotos};
