+++ js/create-thumbnails-list.js
@@ -1,22 +1,22 @@
 const thumbnailsContainer = document.querySelector('.pictures');
 const thumbnailTemplate = document.querySelector('#picture').content.querySelector('.picture');
 
-const createThumbnail = ({url, description, likes, comments}) =>{
-  const thumbnailElement = thumbnailTemplate.cloneNode(true);
-  thumbnailElement.querySelector('.picture__img').src = url;
-  thumbnailElement.querySelector('.picture__img').alt = description;
-  thumbnailElement.querySelector('.picture__likes').textContent = likes;
-  thumbnailElement.querySelector('.picture__comments').textContent = comments.length;
-  return thumbnailElement;
+const createThumbnailElement = ({url, description, likes, comments}) =>{
+  const element = thumbnailTemplate.cloneNode(true);
+  element.querySelector('.picture__img').src = url;
+  element.querySelector('.picture__img').alt = description;
+  element.querySelector('.picture__likes').textContent = likes;
+  element.querySelector('.picture__comments').textContent = comments.length;
+  return element;
 };
 
-const createThumbnailList = (photos) => {
-  const thumbnailsFragment = document.createDocumentFragment();
+const renderThumbnails = (photos) => {
+  const fragment = document.createDocumentFragment();
   photos.forEach((photo) => {
-    const thumbnail = createThumbnail(photo);
-    thumbnailsFragment.append(thumbnail);
+    const thumbnail = createThumbnailElement(photo);
+    fragment.append(thumbnail);
   });
-  thumbnailsContainer.append(thumbnailsFragment);
+  thumbnailsContainer.append(fragment);
 };
 
-export {createThumbnailList};
+export {renderThumbnails};
